chore: split assessment part to `assessmentManager`
